requirements.txt
# Core video processing
ffmpeg-python==0.2.0

# Google Drive API
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-core==2.24.2
googleapis-common-protos==1.70.0

# Clipboard operations
pyperclip==1.8.2

# HTTP requests and API calls
requests==2.32.3

# AI voiceover and transcription
openai-whisper==20240930

# Dependencies (automatically included by pip)
# These are kept for compatibility but not directly used:
# cachetools, certifi, charset-normalizer, decorator, filelock, 
# fsspec, future, httplib2, idna, oauthlib, protobuf, urllib3
